optimize: nested loop retries (#261)
