+++ pkg/rm/init.go
@@ -15,30 +15,13 @@
  * limitations under the License.
  */
 
-package client
+package rm
 
-import (
-	"sync"
-
-	"github.com/seata/seata-go/pkg/remoting/getty"
-)
-
-var onceInitTmClient sync.Once
-
-// InitTmClient init seata tm client
-func initTmClient() {
-	onceInitTmClient.Do(func() {
-		initConfig()
-		initRemoting()
-	})
+// Init init seata client
+func Init() {
+	initRmClient()
 }
 
-// todo
-// initConfig init config processor
-func initConfig() {
-}
-
-// initRemoting init rpc client
-func initRemoting() {
-	getty.InitRpcClient()
+// InitRmClient init seata rm client
+func initRmClient() {
 }
